% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIPCA.R
\name{MRIPCA}
\alias{MRIPCA}
\title{A function for compositional principal component analysis of ROIs across age}
\usage{
MRIPCA(data = NULL, infodir = NULL, Type = 1, Level = 1)
}
\arguments{
\item{data}{use readFileList function to get MRI data frame}

\item{infodir}{path to lookup table of patient info}

\item{Type}{type of MRI sequences(T1-weighted, T2-weighted, etc.), default = 1}

\item{Level}{level of MRI, default = 1}
}
\description{
A function for compositional principal component analysis of ROIs across age
}
